Rename queue types
