+++ src/queue/processors/db/import-user-lists.ts
@@ -8,7 +8,7 @@ import { downloadTextFile } from '../../../misc/download-text-file';
 import { isSelfHost, toPuny } from '../../../misc/convert-host';
 import { DriveFiles, Users, UserLists, UserListJoinings } from '../../../models';
 import { genId } from '../../../misc/gen-id';
-import { DbUserImportJobData } from '../../type';
+import { DbUserImportJobData } from '../../types';
 
 const logger = queueLogger.createSubLogger('import-user-lists');
