+++ src/queue/processors/object-storage/delete-file.ts
@@ -1,6 +1,6 @@
 import * as Bull from 'bull';
 import { deleteObjectStorageFile } from '../../../services/drive/delete-file';
-import { DeleteObjectStorageFileJobData } from '../../type';
+import { DeleteObjectStorageFileJobData } from '../../types';
 
 export default async (job: Bull.Job<DeleteObjectStorageFileJobData>) => {
 	const key: string = job.data.key;
